backend/pyproject.toml
[tool.poetry]
name = "budget-bo-backend"
version = "0.1.0"
description = "SaaS Spending Tracker - Backend API"
authors = ["Budget Bo Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlmodel = "^0.0.22"
asyncpg = "^0.30.0"
redis = "^5.2.0"
celery = "^5.4.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.27.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
cryptography = "^43.0.0"
python-multipart = "^0.0.17"
itsdangerous = "^2.2.0"
duckduckgo-search = "^6.3.0"
woob = "^3.7"
python-levenshtein = "^0.26.0"
ipdb = "^0.13.0"
ipython = "^8.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.27.0"
mypy = "^1.13.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
